Split message log record module
